spice-vdagentd.service: Replace "dbus.target"
dbus.target doesn't exist: instead you should be using dbus.socket

But upon further investigation of spice-vdagentd, it appears that the
only place dbus is used is when trying to talk to logind (or console
kit, which won't be used in tandem with systemd). So really the service
depends on logind, not dbus. So let's represent that in the unit file