spice-vdagent.service: Statically enable
In a previous commit, we configured spice-vdagent.service to only run if
the system socket is running, and the system socket is configured to run
whenever a SPICE virtio channel exists. For the agent to work correctly,
we always want to run the user service whenever the system service is
running.

To that end, we statically enable spice-vdagent.service for any
graphical session. If a SPICE virtio channel exists, the system socket
will exist, and so systemd will start the service. If a channel doesn't
exist, then no socket will exist and systemd will make no attempt to run
the user service.